Add files via upload
In order to use the content of this project the user needs to be logged in to the SQL server before executing the code and make sure using your own personal server name and set it up to the Jason file and change it into your server name from the connection string and then run the code. Before making those changes the code won't run and gives you a connection failure error that required to be fixed.
You have to make sure your connection string matches exactly with your server'a name and the password and username and the name of the database that you use each time connecting to the SQL Server Management.